Remove some Gulp duplication
