+++ src/tasks/styles.js
@@ -1,5 +1,4 @@
-const $ = Elixir.Plugins;
-const config = Elixir.config;
+import combine from './shared/Combine';
 
 /*
  |----------------------------------------------------------------
@@ -13,46 +12,12 @@ const config = Elixir.config;
  */
 
 Elixir.extend('styles', function(styles, output, baseDir) {
-    const paths = prepGulpPaths(styles, baseDir, output);
-
-    new Elixir.Task('styles', function() {
-        return gulpTask.call(this, paths);
-    })
-    .watch(paths.src.path)
-    .ignore(paths.output.path);
+    combine('styles', getPaths(styles, baseDir, output))
 });
-
 
 Elixir.extend('stylesIn', function(baseDir, output) {
-    const paths = prepGulpPaths('**/*.css', baseDir, output);
-
-    new Elixir.Task('stylesIn', function() {
-        return gulpTask.call(this, paths);
-    })
-    .watch(paths.src.path)
-    .ignore(paths.output.path);
+    combine('stylesIn', getPaths('**/*.css', baseDir, output));
 });
-
-
-/**
- * Trigger the Gulp task logic.
- *
- * @param {GulpPaths} paths
- */
-const gulpTask = function(paths) {
-    this.log(paths.src, paths.output);
-
-    return (
-        gulp
-        .src(paths.src.path)
-        .pipe($.if(config.sourcemaps, $.sourcemaps.init()))
-        .pipe($.concat(paths.output.name))
-        .pipe($.if(config.production, require('./shared/CssMinifier')()))
-        .pipe($.if(config.sourcemaps, $.sourcemaps.write('.')))
-        .pipe(gulp.dest(paths.output.baseDir))
-        .pipe(new Elixir.Notification('Stylesheets Merged!'))
-    );
-};
 
 
 /**
@@ -63,8 +28,8 @@ const gulpTask = function(paths) {
  * @param  {string|null}  output
  * @return {GulpPaths}
  */
-const prepGulpPaths = function(src, baseDir, output) {
+const getPaths = function(src, baseDir, output) {
     return new Elixir.GulpPaths()
-        .src(src, baseDir || config.get('assets.css.folder'))
-        .output(output || config.get('public.css.outputFolder'), 'all.css');
+        .src(src, baseDir || Elixir.config.get('assets.css.folder'))
+        .output(output || Elixir.config.get('public.css.outputFolder'), 'all.css');
 };
